implement last view node reg restore ater catchup
