+++ storage/kv_in_memory.py
@@ -64,6 +64,8 @@ class KeyValueStorageInMemory(KeyValueStorage):
                              "should be true")
 
         def filter(key, start, end):
+            if not isinstance(key, str):
+                key = int(key.decode())
             if start and end:
                 return key in range(start, end)
             if start:
@@ -73,7 +75,11 @@ class KeyValueStorageInMemory(KeyValueStorage):
 
         if include_key and include_value:
             if start or end:
-                return {k: v for k, v in self._dict.items() if filter(k, start, end)}
+                filtered_dct = OrderedDict()
+                for k, v in self._dict.items():
+                    if filter(k, start, end):
+                        filtered_dct[k] = v
+                return filtered_dct.items()
             return self._dict.items()
         if include_key:
             if start or end:
